+++ src/input.rs
@@ -25,8 +25,8 @@ pub use host_network::Kind as HostNetworkKind;
 pub use host_network::Model as HostNetworkHtml;
 pub use item::{
     CheckboxItem, ComparisonItem, FileItem, Float64Item, GroupItem, HostNetworkGroupItem,
-    InputItem, NicItem, PasswordItem, PercentageItem, SelectMultipleItem, SelectSingleItem,
-    TagItem, TextItem, Unsigned32Item, VecSelectItem,
+    InputItem, NicItem, PasswordItem, PercentageItem, RadioItem, SelectMultipleItem,
+    SelectSingleItem, TagItem, TextItem, Unsigned32Item, VecSelectItem,
 };
 use strum_macros::{Display, EnumIter, EnumString};
 pub use tag::Model as Tag;